// src/models/TipoCriterioBolao.tsx
export enum TipoCriterioPontuacaoBolao {
    POR_JOGO = 'porJogo',
    EXTRA_1 = 'extra1',
    EXTRA_2 = 'extra2',
    CONVOCACAO_BONUS = 'convocacaoBonus'
}

export enum TipoCriterioPremiacaoBolao {
    LUGAR_GERAL_1 = 'pct1LugarGeral',
    LUGAR_GERAL_2 = 'pct2LugarGeral',
    LUGAR_GERAL_3 = 'pct3LugarGeral',
    LUGAR_GERAL_4 = 'pct4LugarGeral',
    LUGAR_GERAL_5 = 'pct5LugarGeral',
    LUGAR_GERAL_6_10 = 'pct610LugarGeral',
    VENCEDOR_RANK_DIFERENCA_GOLS = 'pctVencedorRankDiferencaGols',
    VENCEDOR_RANK_GOLS = 'pctVencedorRankGols',
    VENCEDOR_RANK_RESULTADO = 'pctVencedorRankResultado',
    VENCEDOR_RANK_PLACAR = 'pctVencedorRankPlacar',
    VENCEDOR_1_FASE = 'pctVencedor1Fase',
    VENCEDOR_2_FASE = 'pctVencedor2Fase',
    LUGAR_PLAYOFF_1 = 'pct1LugarPlayoff',
    LUGAR_PLAYOFF_2 = 'pct2LugarPlayoff',
    LUGAR_PLAYOFF_3 = 'pct3LugarPlayoff',
    LUGAR_GRUPOS_1 = 'pct1LugarGrupos'
}